Start of Cinemachine and sprite atlas.
Unparented the main camera from the player and added a Cinemachine virtual camera. The virtual camera hasn't been changed yet due to issues. The player jitters as they move (both horizontally and vertically), which became noticeable with the new camera. The camera also acts weird when lookahead time was added, jumping forward too much when the player first moves, correcting itself, and then randomly doing jumping forward again. This is all likely to do with how horizontal player movement is currently handled, which is already something that needs to be sorted out.