Moving platforms improvement and a couple other things
-Improved moving platforms a bit.
(Moved the player character's Flip FSM from the main player object to the sprite rendering object parented to it to stop the whole player object stretching horizontally when moving left on the platform. Also adjusted the trigger box so that it's just a thin bar along the top of the platform that doesn't quite reach either end. This solved the issue of the player object sticking to the side of the platform when jumping at the side while the platform is moving; it's only parented to the platform when actually standing on top of the platform.)

-Added a No Input state for the XMove FSM.
(In this state, which is active when the Get Axis (Horizontal) value is around zero, the player object's velocity isn't set each frame like for the Input state. This was implemented while trying to make the player character properly move side to side on platforms while the platform is also in motion (currently the player character will slowly drag itself along strangely, struggling to move from one end to the other). The issue wasn't fixed, but this may be a good feature down the line anyway. Currently there's two versions of the XMove FSM, one with this new implementation. The old one is selected/active, while the version with the No Input state is deactivated but existing.)

-In trying to fix the Shovemeister, I ended up making an unfinished bounce pad.
(The main issue is that the player can jump at the same time force is added, in order to do an unintended super-jump. This should be a fairly easy fix though, such as simply disabling the player character's jump for a moment, which should happen through an action at the start of the state that applies the force, before the force action.

-Slight reshaping of the test scene.